feat: full coverage of response promises
Added support for `fetch.seed` - when the value is set, unless it
matches the current seed, the cache will be wiped. This can be done in
the code or in the devtools console to make repeat testing quick.

Also removes Proxy requirements, sad, and it was cool, but simpler =
better.